+++ src/compas/datastructures/graph/graph.py
@@ -622,12 +622,12 @@ class Graph(Datastructure):
             else:
                 yield key, self.node_attributes(key)
 
-    def nodes_where(self, conditions, data=False):
+    def nodes_where(self, conditions=None, data=False, **kwargs):
         """Get nodes for which a certain condition or set of conditions is true.
 
         Parameters
         ----------
-        conditions : dict
+        conditions : dict, optional
             A set of conditions in the form of key-value pairs.
             The keys should be attribute names. The values can be attribute
             values or ranges of attribute values in the form of min/max pairs.
@@ -641,6 +641,9 @@ class Graph(Datastructure):
             If `data` is True, the next node and its attributes.
 
         """
+        conditions = conditions or {}
+        conditions.update(kwargs)
+
         for key, attr in self.nodes(True):
             is_match = True
 
@@ -740,17 +743,19 @@ class Graph(Datastructure):
                 else:
                     yield u, v
 
-    def edges_where(self, conditions, data=False):
+    def edges_where(self, conditions=None, data=False, **kwargs):
         """Get edges for which a certain condition or set of conditions is true.
 
         Parameters
         ----------
-        conditions : dict
+        conditions : dict, optional
             A set of conditions in the form of key-value pairs.
             The keys should be attribute names. The values can be attribute
             values or ranges of attribute values in the form of min/max pairs.
         data : bool, optional
             If True, yield the edge attributes in addition to the edge identifiers.
+        **kwargs : dict[str, Any], optional
+            Additional conditions provided as named function arguments.
 
         Yields
         ------
@@ -759,6 +764,9 @@ class Graph(Datastructure):
             If `data` is True, the next edge identifier and its attributes as a ((u, v), attr) tuple.
 
         """
+        conditions = conditions or {}
+        conditions.update(kwargs)
+
         for key in self.edges():
             is_match = True